added basic styles and created FormViews Component
