+++ src/styles/global-styles.js
@@ -5,6 +5,7 @@ const GlobalStyle = createGlobalStyle`
     height: 100%;
     width: 100%;
     line-height: 1.4;
+    box-sizing:border-box;
   }
 
   body {
